login user

+++ main.py
@@ -3,27 +3,52 @@ from flask_sqlalchemy import SQLAlchemy
 import os
 from datetime import date, datetime
 
+from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import login_user, LoginManager, current_user, login_required, logout_user
+from flask_security import UserMixin
+import email_validator
 
-admin = 0
+
 
 app = Flask(__name__)
+
+#security
 app.secret_key = "swag"
+login_manager = LoginManager(app)
+
 file_path = os.path.abspath(os.getcwd())+"/database.db"
 app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///" + file_path
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 
 
+
+@login_manager.user_loader
+def get_user(id):
+    return User.query.filter_by(id=id).first()
+
+
+class User(db.Model, UserMixin):
+    id = db.Column(db.Integer, primary_key=True)
+    password = db.Column(db.String(100), nullable=False)
+
+    def __init__(self, password):
+        self.password = generate_password_hash(password)
+
+    def is_active(self):
+        return True
+
+
+
 class Task(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(75), nullable=False, unique=True)
     period = db.Column(db.Integer, nullable=False)
-    description = db.Column(db.String(150), nullable=False, unique=True)
+    description = db.Column(db.String(150), unique=True)
     shift = db.Column(db.String(6), nullable=False)
     next_alert = db.Column(db.DateTime)
-
     done_by = db.Column(db.String(150), nullable=True)
-    #done_date = db.Column(db.DateTime)
+
 
 
     def __init__(self, name, period, description, shift, next_alert):
@@ -33,28 +58,35 @@ class Task(db.Model):
         self.shift = shift
         self.next_alert = next_alert
 
-    def is_active(self):
-        return True
 
-class done_Task:
+class Done_Task(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
     done_date = db.Column(db.DateTime(timezone=True))
     person_name = db.Column(db.String(150), nullable=False)
-    task_name = db.Column(db.String(75), nullable=False, unique=True)
+    task_name = db.Column(db.String(75), nullable=False)
 
     def __init__(self, person_date, task_name):
         self.done_date = datetime.datetime.today().replace(microsecond=0)
         self.person_name = person_date
         self.task_name = task_name
 
-
 db.create_all()
+
+main_admin = User('admin')
+db.session.add(main_admin)
+db.session.commit()
+
+
+
 @app.route('/', methods=['POST', 'GET'])
 def home():
     if request.method == 'POST':
-        password = request.form.get('admin')
-        if password == 'admin':
-            global admin
-            admin = 1
+        password = request.form.get('password')
+        
+        test = User.query.filter_by(id=1).first()
+
+        if test and check_password_hash(test.password, password):
+            login_user(test)
             return redirect(url_for('admin'))
         else:
             flash("Wrong password for admin!", 'error')
@@ -68,26 +100,29 @@ def home():
 
 
 @app.route('/admin', methods=['POST', 'GET'])
+@login_required
 def admin():
-    if admin == 1:
-        today_d = datetime.now().replace(microsecond=0, hour=0, second=0, minute=0)
-        return render_template("admin.html", tasks=Task.query.filter_by(next_alert=today_d))
-    else:
-        return redirect(url_for('home'))
+    today_d = datetime.now().replace(microsecond=0, hour=0, second=0, minute=0)
+    return render_template("admin.html", tasks=Task.query.filter_by(next_alert=today_d))
+
 
 
 @app.route('/select', methods=['POST', 'GET'])
+@login_required
 def select():
     return render_template("select.html")
 
 @app.route('/manage', methods=['POST', 'GET'])
+@login_required
 def manage():
     if request.method == 'POST':
         name = request.form.get('name')
         des = request.form.get('description')
         shift = request.form.get('shift')
+
         if request.form['first_a'] and request.form['second_a']:
             start_date = datetime.strptime(request.form['first_a'], '%Y-%m-%d').date()
+            print(start_date)
             second_date = datetime.strptime(request.form['second_a'], '%Y-%m-%d').date()
         else:
             flash("Your task doesn't have some of his alerts!", 'error')
@@ -121,8 +156,9 @@ def manage():
 
 
 @app.route('/edit/<my_task_id>', methods=['POST', 'GET'])
+@login_required
 def edit(my_task_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and current_user.get_id() == 1:
         task = Task.query.filter_by(id=my_task_id).first()
 
         task_name_new = request.form.get('task_name_edit')
@@ -131,6 +167,7 @@ def edit(my_task_id):
         if 'first_alert_edit' in request.form and 'second_alert_edit' in request.form:
             start_date = datetime.strptime(request.form['first_alert_edit'], '%Y-%m-%d').date()
             second_date = datetime.strptime(request.form['second_alert_edit'], '%Y-%m-%d').date()
+
             if second_date < start_date:
                 flash('Your period is negative!', 'error')
                 return redirect(url_for('manage'))
@@ -140,6 +177,7 @@ def edit(my_task_id):
                 return redirect(url_for('manage'))
 
             period = second_date - start_date
+            period = period.days
             task.period = period
             task.next_alert = start_date
 
@@ -172,9 +210,9 @@ def edit(my_task_id):
 
 
 @app.route('/delete/<task_id>', methods=['POST', 'GET'])
+@login_required
 def delete(task_id):
-    if request.method == 'POST':
-        print(task_id)
+    if request.method == 'POST' and current_user.get_id() == 1:
         Task.query.filter_by(id=task_id).delete()
         db.session.commit()
 
@@ -182,12 +220,13 @@ def delete(task_id):
     else:
         return redirect(url_for('manage'))
 
+
 @app.route('/send/<task_id>', methods=['POST', 'GET'])
 def send(task_id):
     if request.method == 'POST':
         task = Task.query.filter_by(id=task_id).first()
         person_name = request.form.get('person_name')
-        done_task = done_Task(date.today(), person_name, task.name)
+        done_task = Done_Task(person_name, task.name)
 
         db.session.add(done_task)
         db.session.commit()
@@ -196,6 +235,12 @@ def send(task_id):
     else:
         return redirect(url_for('home'))
 
+
+@app.route('/logout', methods=['POST', 'GET'])
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('home'))
 
 if __name__ == '__main__':
     app.run(debug=True)
